+++ source/others/f16_to_extF80M.cpp
@@ -83,7 +83,7 @@ void f16_to_extF80M(float16_t a, extFloat80_t *zPtr)
         }
     }
 
-    uiZ64 = packToExtF80UI64(sign, exp + 0x3FF0);
+    uiZ64 = packToExtF80UI64(sign, static_cast<uint16_t>(exp + 0x3FF0));
     uiZ32 = 0x80000000 | (uint32_t)frac << 21;
 uiZ:
     zSPtr->signExp = uiZ64;
