+++ source/others/f32_to_extF80M.cpp
@@ -79,7 +79,7 @@ f32_to_extF80M(float32_t a, extFloat80_t *zPtr)
                 frac = normExpSig.sig;
             }
         }
-        zSPtr->signExp = packToExtF80UI64(sign, exp + 0x3F80);
+        zSPtr->signExp = packToExtF80UI64(sign, static_cast<uint16_t>(exp + 0x3F80));
         zSPtr->signif = (uint64_t)(0x80000000 | (uint32_t)frac << 8) << 32;
     }
 }
