+++ source/others/s_normSubnormalF16Sig.cpp
@@ -37,12 +37,12 @@ SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 #include "internals.hpp"
 
 exp8_sig16
-softfloat_normSubnormalF16Sig( uint16_t sig )
+softfloat_normSubnormalF16Sig(uint16_t sig)
 {
-    int8_t const shiftDist = softfloat_countLeadingZeros16( sig ) - 5;
+    int8_t const shiftDist = softfloat_countLeadingZeros16(sig) - 5;
     exp8_sig16 z;
     z.exp = 1 - shiftDist;
-    z.sig = sig<<shiftDist;
+    z.sig = static_cast<uint16_t>(sig << shiftDist);
     return z;
 }
